pyproject.toml
[tool.poetry]
name = "agent-army"
version = "0.2.0"
description = "AI-Team Multi-Agent Lead-Generation System for B2B Sales Automation"
authors = ["Fabiano <fabiano@frascati-systems.ch>"]
readme = "README.md"
packages = [{include = "agent_army", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
asyncio-mqtt = "^0.16.2"
aiohttp = ">=3.9.1"
aiosqlite = ">=0.19.0"
sqlalchemy = {extras = ["asyncio"], version = ">=2.0.23"}
pydantic = ">=2.5.2"
pydantic-settings = ">=2.1.0"
rich = ">=13.7.0"
loguru = ">=0.7.2"
click = ">=8.1.7"
pyyaml = ">=6.0.1"
beautifulsoup4 = ">=4.12.2"
lxml = ">=4.9.4"
aiosmtplib = ">=3.0.1"
aioimaplib = ">=1.0.1"
tenacity = ">=8.2.3"
websockets = ">=12.0"
httpx = ">=0.25.2"
python-dateutil = ">=2.8.2"
jinja2 = ">=3.1.2"
anthropic = ">=0.76.0"
fastapi = ">=0.115.0"
uvicorn = {extras = ["standard"], version = ">=0.34.0"}
python-multipart = ">=0.0.18"
playwright = ">=1.49.0"
crawl4ai = ">=0.6.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.4.3"
pytest-asyncio = ">=0.23.2"
pytest-cov = ">=4.1.0"
black = ">=23.12.1"
mypy = ">=1.7.1"
ruff = ">=0.1.8"

[tool.poetry.scripts]
agent-army = "agent_army.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
